requirements.txt
Pillow
qrcode
reportlab